types and typos

+++ example.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <assert.h>
 #include "rusty_enum.h"
 
 RUSTY_ENUM(IpAddr, V4, V6)
-RUSTY_ENUM_FIELDS(V4, int, int, int, int)
+RUSTY_ENUM_FIELDS(V4, uint8_t, uint8_t, uint8_t, uint8_t)
 RUSTY_ENUM_FIELDS(V6, const char *)
 RUSTY_ENUM_END(IpAddr); // neovim _really_ hates it when there is no semicolon here for some reason
 
@@ -13,8 +14,8 @@ int main()
 
     switch (addr.kind) {
         case V4: {
-            int a = RUSTY_ENUM_GET(addr, V4, 0);
-            int b = addr.V4._0;
+            uint8_t a = RUSTY_ENUM_GET(addr, V4, 0);
+            uint8_t b = addr.V4._0;
             assert(a == b);
             break;
         }
